Put site.php in the back (like in epfl-accred)
This ensures that all classes will be defined before we invoke their
singletons.

The reason it was working before — or more precisely, appeared to — is
a bit arcane. PHP tries to compile the whole file the first time it
sees it (opens it), which results in Controller::getInstance() being
callable even before control reaches the require_once(".../site.php")
line (contrary to what a "naive" reading of the control flow would
suggest). However, this early compile can silently fail (and be
retried later) if any dependencies aren't available - Here, the
\EPFL\SettingsBase class that Settings extends. Bottom line, the
following stack trace appeared in the logs iff epfl-tequila was the
only EPFL plug-in currently active:

[Thu Nov 23 18:28:05.187920 2017] [:error] [pid 641] [client 172.19.0.1:49050] PHP Fatal error:  Uncaught Error: Class 'EPFL\\Tequila\\Settings' not found in /srv/test/localhost/htdocs/sti/wp-content/plugins/epfl-tequila/EPFL-Tequila.php:53
Stack trace:
#0 /srv/test/localhost/htdocs/sti/wp-content/plugins/epfl-tequila/EPFL-Tequila.php(46): EPFL\\Tequila\\Controller->__construct()
#1 /srv/test/localhost/htdocs/sti/wp-content/plugins/epfl-tequila/site.php(12): EPFL\\Tequila\\Controller::getInstance()
#2 /srv/test/localhost/htdocs/sti/wp-content/plugins/epfl-tequila/EPFL-Tequila.php(20): require_once('/srv/test/local...')
#3 /srv/test/localhost/htdocs/sti/wp-settings.php(305): include_once('/srv/test/local...')
#4 /srv/test/localhost/htdocs/sti/wp-config.php(76): require_once('/srv/test/local...')
#5 /srv/test/localhost/htdocs/sti/wp-load.php(37): require_once('/srv/test/local...')
#6 /srv/test/localhost/htdocs/sti/wp-admin/admin.php(31): require_once('/srv/test/local...')
#7 /srv/test/localhost/htdocs/sti/wp-admin/plugins.php(10): require_once('/srv/test/local...')
#8 {main}
  thrown in /srv/test/localhost/htdocs/sti/wp-content/plugins/epfl-tequila/EPFL-Tequila.php on line 53, referer: https://localhost:444/sti/wp-admin/plugins.php?plugin_status=all&paged=1&s